Created a Tag page. And a user page. These are are the same format as ForYou.php but users can focus on specific tags and view an authors profile page. Updated the links on these three pages so the author, and tags redirect to these new pages.
Edit-article.php/create-article.php
Updated the forms so it looks professional. The suggestion dropdowns are now attached to the input text box so it moves with the page. I've also added a js handler so if the user clicks off the page whilst the form has elements filled in, an alert is issued to the user.
Fixed the bug where the form would reset if not all the elements were filled in correctly.

Updated article.php to connet the links to the tag and author pages.

Settings.php has been updated to account for the new links. The pages still need to be built but the wireframes are in place.